requirements.txt
# OCR2LATEX 项目依赖

# 核心依赖
opencv-python>=4.8.0          # 图像处理
Pillow>=10.0.0                # 图像操作和增强
numpy>=1.24.0                 # 数值计算
requests>=2.31.0              # HTTP客户端

# 可选依赖（用于PDF支持）
PyMuPDF>=1.23.0               # PDF处理（可选）

# 开发和测试依赖（可选）
pytest>=7.4.0                # 单元测试
pytest-cov>=4.1.0            # 测试覆盖率

